+++ src/routes/chatbot.py
@@ -1,8 +1,10 @@
 from flask import Blueprint
 
+from services.chatbot import chat as cbchat
+
 chatbot_bp = Blueprint('chatbot_bp', __name__)
 
 
 @chatbot_bp.route('/v1/chatbot/chat', methods=['POST'])
 def chat():
-    return {'response': "Hardcoded response..."}, 200
+    return {'response': cbchat.chat()}, 200
