+++ src/services/chatbot/chat.py
@@ -1,0 +1,2 @@
+def chat(*args, **kwargs):
+    return "Hardcoded response..."
